app/token, restful: use RSA key types for key arguments
Instead of passing byte arrays for signing key and verifying key for
various token functions, use rsa.PrivateKey and rsa.PublicKey. This has
the benefit of better type checking and it avoids the need to re-parse
the PEM format on each call to ParseToken or GenerateToken (the PEM file
is parsed only once during TokenSetUp).

Moreover, knowledge of IP address parsing/cleanup is removed from the
token package. It is the responsiblity of the caller to pass the proper
clean IP address (without source port) in the calls to GenerateToken and
ParseToken.